refactor(*): refactor as per feedback
