+++ honeybee_vtk/window.py
@@ -1,18 +1,24 @@
 """A vtk render window."""
 
-from __future__ import annotations
 import vtk
-from typing import List, Tuple
+import warnings
+import pathlib
+from typing import List
+from .actor import Actor
+from .camera import Camera
+from .types import ImageTypes
 
 
 class Window:
-    def __init__(self, actors=None, camera=None):
+    def __init__(self, actors: List[Actor] = None, camera: Camera = None):
         self._actors = actors
         self._camera = camera
+        self._interactor = None
+        self._window = None
+        self._renderer = None
+        self._create_window()
 
-    def create_window(self) -> Tuple[
-            vtk.vtkRenderWindowInteractor, vtk.vtkRenderWindow, vtk.vtkRenderer]:
-
+    def _create_window(self) -> None:
         """Create a rendering window with a single renderer and an interactor.
 
         The objects are embedded inside each other. interactor is the outmost layer.
@@ -20,15 +26,15 @@ class Window:
         This method returns a tuple of a window_interactor, a render_window, and a
         renderer.
         """
-        # if camera is not provided use the first camera in the list of cameras setup
-        # in the scene
         if not self._camera:
-            
+            raise ValueError(
+                'A camera is needed to create a render window.'
+            )
 
         # Setting renderer, render window, and interactor
         renderer = vtk.vtkRenderer()
 
-        # Add actors to the window if model(actor) has arrived at the scene
+        # Add actors to the window
         if self._actors:
             for actor in self._actors.values():
                 renderer.AddActor(actor.to_vtk())
@@ -48,11 +54,172 @@ class Window:
         renderer.TwoSidedLightingOn()
 
         # Assign camera to the renderer
-        if camera.type == 'v':
-            renderer.SetActiveCamera(camera.to_vtk())
+        if self._camera.type == 'v':
+            renderer.SetActiveCamera(self._camera.to_vtk())
         else:
             bounds = Actor.get_bounds(self.actors.values())
-            renderer.SetActiveCamera(camera.to_vtk(bounds=bounds))
+            renderer.SetActiveCamera(self._camera.to_vtk(bounds=bounds))
 
-        # return the objects - the order is from outside to inside
-        return interactor, window, renderer
+        # the order is from outside to inside
+        self._interactor, self._window, self._renderer = (interactor, window, renderer)
+
+    def export_gltf(self, folder: str, name: str) -> str:
+        """Export a scene to a glTF file.
+
+        Args:
+            folder: A valid path to where you'd like to write the gltf file.
+            name: Name of the gltf file as a text string.
+
+        Returns:
+            A text string representing the path to the gltf file.
+        """
+        # TODO: Find out why model's displaymode is not applied
+        # TODO: Find out if the axis should be rotated to work with gltf viewers
+        # TODO: Find a viewer for gltf files. The up axis of f3d viewer is Y axis.
+
+        gltf_file_path = pathlib.Path(folder, f'{name}.gltf')
+        exporter = vtk.vtkGLTFExporter()
+        exporter.SaveNormalOn()
+        exporter.InlineDataOn()
+        exporter.SetFileName(gltf_file_path.as_posix())
+        exporter.SetActiveRenderer(self._renderer)
+        exporter.SetRenderWindow(self._window)
+        exporter.Modified()
+        exporter.Write()
+        return gltf_file_path.as_posix()
+
+    @staticmethod
+    def _get_image_writer(image_type: ImageTypes) -> None:
+        """Get vtk image writer for each image type."""
+        if image_type == ImageTypes.png:
+            writer = vtk.vtkPNGWriter()
+        elif image_type == ImageTypes.bmp:
+            writer = vtk.vtkBMPWriter()
+        elif image_type == ImageTypes.jpg:
+            writer = vtk.vtkJPEGWriter()
+        elif image_type == ImageTypes.pnm:
+            writer = vtk.vtkPNMWriter()
+        elif image_type == ImageTypes.ps:
+            writer = vtk.vtkPostScriptWriter()
+        elif image_type == ImageTypes.tiff:
+            writer = vtk.vtkTIFFWriter()
+        else:
+            raise ValueError(f'Invalid image type: {image_type}')
+        return writer
+
+    def export_image(
+            self, folder: str, name: str, image_type: ImageTypes = ImageTypes.png, *,
+            image_scale: int = 1, image_width: int = 2200, image_height: int = 2000,
+            color_range: vtk.vtkLookupTable = None, rgba: bool = False,
+            show: bool = False) -> str:
+        """Export the scene as an image.
+
+        Reference: https://kitware.github.io/vtk-examples/site/Python/IO/ImageWriter/
+        This method is able to export an image in '.png', '.jpg', '.ps', '.tiff', '.bmp',
+        and '.pnm' formats.
+
+        Args:
+            folder: A valid path to where you'd like to write the image.
+            name: Name of the image as a text string.
+            image_type: An ImageType object.
+            image_scale: An integer value as a scale factor. Defaults to 1.
+            image_width: An integer value that sets the width of image in pixels.
+                Defaults to 2200.
+            image_height: An integer value that sets the height of image in pixels.
+                Defaults to 2000.
+            color_range: A vtk lookup table object which can be obtained
+                from the color_range mehtod of the DataFieldInfo object.
+                Defaults to None.
+            rgba: A boolean value to set the type of buffer. A True value sets
+                an the background color to black. A False value uses the Scene object's
+                background color. Defaults to False.
+            show: A boolean value to decide if the the render window should pop up.
+                Defaults to False.
+
+        Returns:
+            A text string representing the path to the image.
+        """
+
+        if color_range:
+            scalar_bar = vtk.vtkScalarBarActor()
+            scalar_bar.SetOrientationToHorizontal()
+            scalar_bar.SetLookupTable(color_range)
+
+            # create the scalar_bar_widget
+            legend = vtk.vtkScalarBarWidget()
+            legend.SetInteractor(self._interactor)
+            legend.SetScalarBarActor(scalar_bar)
+            legend.On()
+
+        # render window
+        if not show:
+            self._window.OffScreenRenderingOn()
+            self._window.SetSize(image_width, image_height)
+            self._window.Render()
+        else:
+            self._window.SetSize(image_width, image_height)
+            self._window.Render()
+            self._interactor.Initialize()
+            self._interactor.Start()
+
+        image_path = pathlib.Path(folder, f'{name}.{image_type.value}')
+        writer = self._get_image_writer(image_type)
+        if image_type == ImageTypes.jpg:
+            writer.SetQuality(100)  # image quality
+
+        window_to_image_filter = vtk.vtkWindowToImageFilter()
+        window_to_image_filter.SetInput(self._window)
+        window_to_image_filter.SetScale(image_scale)
+
+        # rgba is not supported for postscript image type
+        if rgba and image_type != ImageTypes.ps:
+            window_to_image_filter.SetInputBufferTypeToRGBA()
+        else:
+            window_to_image_filter.SetInputBufferTypeToRGB()
+            # Read from the front buffer.
+            window_to_image_filter.ReadFrontBufferOff()
+            window_to_image_filter.Update()
+
+        writer.SetFileName(image_path.as_posix())
+        writer.SetInputConnection(window_to_image_filter.GetOutputPort())
+        writer.Write()
+
+        if color_range:
+            legend.Off()
+
+        return image_path.as_posix()
+
+    # def export_images(
+    #         self, folder: str, name: str, image_type: ImageTypes = ImageTypes.png, *,
+    #         image_scale: int = 1, image_width: int = 2200, image_height: int = 2000,
+    #         color_range: vtk.vtkLookupTable = None, rgba: bool = False) -> Tuple[str]:
+    #     """Export all the cameras setup in a scene as images.
+
+    #     This method calls export_image method under the hood.
+
+    #     Args:
+    #         folder: A valid path to where you'd like to write the images.
+    #         name: A text string that will be used as a name for the images.
+    #         image_type: An ImageType object.
+    #         image_scale: An integer value as a scale factor. Defaults to 1.
+    #         image_width: An integer value that sets the width of image in pixels.
+    #         image_height: An integer value that sets the height of image in pixels.
+    #         color_range: A vtk lookup table object which can be obtained
+    #             from the color_range mehtod of the DataFieldInfo object.
+    #             Defaults to None.
+    #         rgba: A boolean value to set the type of buffer. A True value sets
+    #             an the background color to black. A False value uses the Scene object's
+    #             background color. Defaults to False.
+
+    #     Returns:
+    #         A tuple of paths to exported images.
+    #     """
+    #     file_paths = []
+
+    #     for count, camera in enumerate(self._cameras):
+    #         file_paths.append(
+    #             self.export_image(folder=folder, name=name + '_' + str(count),
+    #                               image_type=image_type, image_scale=image_scale,
+    #                               image_width=image_width, image_height=image_height,
+    #                               color_range=color_range, rgba=rgba))
+    #     return tuple(file_paths)
